% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_header.R
\name{clean_header}
\alias{clean_header}
\title{Clean multiline column headers}
\usage{
clean_header(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
dataframe
}
\description{
All destatis table come with multiline headers.
This function tries to get them in a machine readable form by building new column names.
It takes into account that all column names must be unique. So duplicate column names will be marked with the index.
Non-character parts like `(%)` or `=100` will be replaced.
}
