% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_copyright.R
\name{delete_copyright}
\alias{delete_copyright}
\title{Delete copyright and metadata information}
\usage{
delete_copyright(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
dataframe
}
\description{
Copyright information and metadata is positioned at the end of the file.
It will be detected and replaced by looking for the first row of the last/right column that is empty.
}
