% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/destatiscleanr.R
\name{destatiscleanr}
\alias{destatiscleanr}
\title{Import and clean data from official German statistical offices to jump-start the data analysis.}
\usage{
destatiscleanr(file)
}
\arguments{
\item{file}{path to the csv file, that should be cleaned}
}
\value{
dataframe
}
\description{
This is the main function of the package destatiscleanr.
It calls four functions to read the file, delete copyright and metadata information,
cleans the column headers and converts certain columns to numeric.
}
\examples{
\dontrun{
destatiscleanr(file = "path/to/destatis_table.csv")
}
}
