% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file.R
\name{read_file}
\alias{read_file}
\title{This functions reads the csv file by using German decimal marks}
\usage{
read_file(file)
}
\arguments{
\item{file}{path to the csv file, that should be cleaned}
}
\value{
dataframe
}
\description{
This functions reads the csv file by using German decimal marks
}
